change pressure api to doubles fixed some imports fixed gitignore updated mcmod.info removed old version of waila

+++ src/main/java/com/cout970/magneticraft/api/pressure/IPressureConductor.java
@@ -32,8 +32,8 @@ public interface IPressureConductor extends IConnectable{
 	 * moles of gas in the conductor
 	 * @return
 	 */
-	public long getMoles();
-	public void setMoles(long moles);
+	public double getMoles();
+	public void setMoles(double moles);
 	
 	/**
 	 * temperature of the gas in kelvin
